pyproject.toml
[project]
name = "TPTK"
version = "0.0.8"
authors = [{ name = "Gaurav Jaiswal", email = "jaiswalgaurav863@gmail.com" }]
description = "A Python package for automating text preprocessing tasks."
readme = "README.md"
requires-python = ">=3.8"
license = { text = "MIT" }
keywords = ["text preprocessing", "NLP", "text cleaning"]
dependencies = ["nltk", "pyspellchecker", "pandas"]
